Add `umd` & `es` targets
Rename demo page

Add start script

Improve component loading
