+++ README.md
@@ -1,3 +1,10 @@
 # &lt;file-link&gt;
 
 `<file-link>` web component
+
+## How to
+
+### Demo
+```bash
+npm start
+```
