+++ package.json
@@ -1,16 +1,30 @@
 {
   "name": "@netology-group/wc-file-link",
-  "version": "0.1.3",
+  "version": "0.1.4",
   "description": "`file-link` web component",
-  "main": "src/index.js",
-  "scripts": {
-    "build": "webpack --mode production",
-    "build:dev": "webpack --mode development"
-  },
   "keywords": [],
-  "author": "Aleksey Konstantinov <a.konstantinov@netology-group.ru>",
   "license": "MIT",
+  "author": "Aleksey Konstantinov <a.konstantinov@netology-group.ru>",
+  "files": [
+    "dist",
+    "es"
+  ],
+  "main": "dist/file-link.js",
+  "module": "es/file-list.js",
+  "directories": {
+    "lib": "src"
+  },
+  "scripts": {
+    "prebuild": "rm -rf dist es",
+    "build": "npm run build:es && npm run build:umd:min",
+    "build:dev": "cross-env BABEL_ENV=umd webpack --mode development",
+    "build:es": "cross-env BABEL_ENV=es babel --copy-files src -d es",
+    "build:umd:min": "cross-env BABEL_ENV=umd webpack --mode production",
+    "demo": "http-server",
+    "start": "npm run demo"
+  },
   "dependencies": {
+    "@netology-group/wc-utils": "^0.3.2",
     "@polymer/lit-element": "~0.5.2",
     "lit-html": "~0.10.2",
     "mime": "~2.3.1",
@@ -20,10 +34,13 @@
     "autoprefixer": "~9.0.0",
     "babel-core": "~6.26.3",
     "babel-loader": "~7.1.4",
+    "babel-plugin-annotate-pure-calls": "^0.4.0",
     "babel-preset-env": "~1.7.0",
+    "cross-env": "^5.2.0",
     "cssnano": "~4.0.3",
     "postcss-loader": "~2.1.6",
     "postcss-url": "~7.3.2",
+    "uglifyjs-webpack-plugin": "^2.0.1",
     "webpack": "~4.12.0",
     "webpack-cli": "~3.0.6"
   }
